requirements.txt
Pillow>=9.0.0
PyGObject>=3.40.0
